refactored helpers into more meanginful file names
