+++ mobile-browser-based-version/server/src/helpers/authenticator.ts
