+++ mobile-browser-based-version/server/src/helpers/tensor_serializer.ts
@@ -11,7 +11,7 @@ async function serializeTensor (tensor) {
   }
 }
 
-function deserializeTensor (dict) {
+export function deserializeTensor (dict) {
   const { data, shape, dtype } = dict.$tensor
   return tf.tensor(data, shape, dtype).clone() // doesn't copy (maybe depending on runtime)!
 }
@@ -27,30 +27,6 @@ async function serializeVariable (variable) {
 
 export async function serializeWeights (weights) {
   return await Promise.all(weights.map(serializeVariable))
-}
-
-export async function averageWeights (roundSerializedWeights) {
-  const firstWeights = roundSerializedWeights[0]
-  const otherWeights = roundSerializedWeights.slice(1)
-  const resultWeights = []
-  firstWeights.forEach((weight, idx) => {
-    let tensorSum = deserializeTensor(weight.$variable.val)
-    otherWeights.forEach((serializedWeights) => {
-      const serializedWeight = serializedWeights[idx].$variable
-      const tensor = deserializeTensor(serializedWeight.val)
-      tensorSum = tensor.add(tensorSum)
-      tensor.dispose()
-    })
-    const val = tensorSum.div(roundSerializedWeights.length)
-    const newWeight = {
-      name: weight.$variable.name,
-      val: val
-    }
-    resultWeights.push(newWeight)
-    tensorSum.dispose()
-  })
-  const serializedWeights = serializeWeights(resultWeights)
-  return serializedWeights
 }
 
 export function assignWeightsToModel (model, serializedWeights) {
