+++ mobile-browser-based-version/src/components/TaskList.vue
@@ -52,10 +52,8 @@ import CustomButton from './simple/CustomButton.vue'
 import _ from 'lodash'
 import { defineComponent } from 'vue'
 import { mapMutations } from 'vuex'
-import {
-  createTaskClass,
-  loadTasks
-} from '../helpers/task_definition/helper'
+import { createTaskClass } from '../helpers/task_definition/task_builder'
+import { loadTasks } from '../helpers/task_definition/tasks_io'
 
 export default defineComponent({
   name: 'task-list',
