+++ mobile-browser-based-version/src/components/building_frames/DescriptionFrame.vue
@@ -1,4 +1,3 @@
-import { serializeWeights } from '../helpers/tfjs_helpers.js';
 <template>
   <div>
   <a id="overview-target">
@@ -164,7 +163,7 @@ import { serializeWeights } from '../helpers/tfjs_helpers.js';
 </template>
 
 <script>
-import * as memory from '../../helpers/memory/helpers'
+import * as memory from '../../helpers/memory/model_io'
 import CustomButton from '../simple/CustomButton.vue'
 import Tasks from '../../assets/svg/Tasks.vue'
 import Model from '../../assets/svg/Model.vue'
