+++ mobile-browser-based-version/src/helpers/logging/console_logger.ts
